Add Arpack/TSVD weakdeps extension
Move Arpack and TSVD out of mandatory deps into weakdeps and add an extension (ext/LinearOperatorsArpackTSVDExt.jl) that provides estimate_opnorm methods when those packages are available. The new extension implements opnorm estimation via ARPACK eigs/svds and TSVD (with small-matrix dense fallbacks, retry/NCV growth logic and ARPACK error handling). The original implementation in src/utilities.jl was removed (left as a stub) and tests/Project.toml/tests updated to load TSVD and Arpack for the test suite.
